# feat: parallelize JSONL file parsing for 1.6x performance improvement