## Summary

- Implements parallel JSONL file parsing using goroutines and channels for significant performance improvement
- Adds comprehensive benchmark tests to measure performance gains
- Maintains backward compatibility and identical error handling behavior

## Performance Results

**Benchmark comparison (go test -bench=BenchmarkParseFiles ./internal/claude):**

- **Sequential**: 11,678,763 ns/op (~11.7ms per operation)
- **Parallel**: 7,345,007 ns/op (~7.3ms per operation)
- **Improvement**: **1.6x faster** (37% performance gain)

## Technical Implementation

- **Worker Pool**: Uses `runtime.GOMAXPROCS(0)` to determine optimal worker count
- **Channels**: File paths distributed via channels, results collected concurrently
- **Error Handling**: Maintains same error behavior - skips invalid files and continues
- **Synchronization**: Uses `sync.WaitGroup` for proper goroutine coordination

## Testing

- Added `BenchmarkParseFilesSequential` and `BenchmarkParseFilesParallel` tests
- All existing unit tests pass
- Manual testing confirms identical output and behavior

## Test Plan

- [x] Unit tests pass
- [x] Benchmark tests show performance improvement
- [x] Manual testing with real data confirms identical output
- [x] Error handling behavior unchanged
- [x] No breaking changes to API

🤖 Generated with [Claude Code](https://claude.ai/code)